pyproject.toml
[project]
name = "resumidor"
version = "0.1.0"
description = ""
authors = [
    {name = "Protons"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "streamlit (>=1.44.0,<2.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "openai (>=1.69.0,<2.0.0)",
    "tenacity (>=9.0.0,<10.0.0)",
    "pydantic (>=2.11.1,<3.0.0)",
    "langchain (>=0.3.22,<0.4.0)",
    "langchain-openai (>=0.3.11,<0.4.0)",
    "pymongo (>=4.11.3,<5.0.0)",
    "qdrant-client (>=1.13.3,<2.0.0)",
    "pymupdf4llm (>=0.0.19,<0.0.20)",
    "youtube-transcript-api (>=1.0.3,<2.0.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "markdownify (>=1.1.0,<2.0.0)",
    "arxiv (>=2.1.3,<3.0.0)",
    "googlesearch-python (>=1.3.0,<2.0.0)",
    "huggingface-hub (>=0.30.1,<0.31.0)",
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "transformers (>=4.50.3,<5.0.0)",
    "torch (>=2.6.0,<3.0.0)",
    "einops (>=0.8.1,<0.9.0)",
    "tiktoken (>=0.9.0,<0.10.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
langchain-community = "^0.3.20"
pytest = "^8.3.5"

